backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pandas==2.2.2
openpyxl==3.1.2
Pillow==10.3.0
PyMySQL==1.1.1
cryptography==42.0.5
python-dotenv==1.0.1
bcrypt==4.1.3
SQLAlchemy==2.0.30
Werkzeug==3.0.3
